python-backend/requirements.txt
# Core
fastapi==0.115.0
uvicorn[standard]==0.30.0
python-dotenv==1.0.0
python-multipart==0.0.9

# LangChain
langchain==0.3.0
langchain-google-genai==2.0.0
langchain-pinecone==0.2.0
langchain-community==0.3.0

# Embeddings & LLM
google-generativeai==0.8.0

# Document Loaders
pypdf==5.0.0
python-pptx==1.0.0
python-docx==1.1.0

# Vector Store
pinecone-client==5.0.0

# Web Scraping
beautifulsoup4==4.12.0
requests==2.32.0

# CORS
sse-starlette==2.0.0
